requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.5.7
python-telegram-bot==13.11
pytz==2023.3
six==1.16.0
tornado==6.3.2
tzdata==2023.3
tzlocal==5.0.1
